requirements.txt
pyyaml==5.4.1
flask==1.1.2
pymongo==3.11.3
dnspython==2.0.0
waitress==1.4.4
paste==3.5.0